+++ kubejs/server_scripts/recipes/mekanism/metallurgic_infusing.js
@@ -7,6 +7,24 @@ ServerEvents.recipes((event) => {
             item_input: { tag: 'c:foods/raw_meat', count: 1 },
             chemical_input: { amount: 10, tag: 'mekanism:bio' },
             id: `${id_prefix}rotten_flesh`
+        },
+        {
+            output: { id: 'minecraft:golden_carrot', count: 1 },
+            item_input: { item: 'minecraft:carrot', count: 1 },
+            chemical_input: { amount: 10, tag: 'mekanism:gold' },
+            id: `${id_prefix}golden_carrot`
+        },
+        {
+            output: { id: 'minecraft:glistering_melon_slice', count: 1 },
+            item_input: { item: 'minecraft:melon_slice', count: 1 },
+            chemical_input: { amount: 10, tag: 'mekanism:gold' },
+            id: `${id_prefix}glistering_melon_slice`
+        },
+        {
+            output: { id: 'minecraft:golden_apple', count: 1 },
+            item_input: { item: 'minecraft:apple', count: 1 },
+            chemical_input: { amount: 80, tag: 'mekanism:gold' },
+            id: `${id_prefix}golden_apple`
         }
     ];